finished db implementation

+++ db_ops.py
@@ -49,8 +49,8 @@ try:
         # Define the SQL query to create the 'baby' table
         create_baby_table_query = '''
         CREATE TABLE IF NOT EXISTS baby (
-            baby_id INT AUTO_INCREMENT,
-            user_id INT,
+            baby_id BIGINT AUTO_INCREMENT,
+            user_id BIGINT,
             name VARCHAR(50),
             date_of_birth DATE,
             created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
@@ -65,8 +65,8 @@ try:
         # Define the SQL query to create the 'height' table
         create_height_table_query = '''
         CREATE TABLE IF NOT EXISTS height (
-            log_id INT AUTO_INCREMENT,
-            baby_id INT,
+            log_id BIGINT AUTO_INCREMENT,
+            baby_id BIGINT,
             height FLOAT,
             log_date TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
             PRIMARY KEY (log_id)
@@ -80,8 +80,8 @@ try:
         # Define the SQL query to create the 'weight' table
         create_weight_table_query = '''
         CREATE TABLE IF NOT EXISTS weight (
-            log_id INT AUTO_INCREMENT,
-            baby_id INT,
+            log_id BIGINT AUTO_INCREMENT,
+            baby_id BIGINT,
             weight FLOAT,
             log_date TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
             PRIMARY KEY (log_id)
@@ -95,14 +95,13 @@ try:
         # Define the SQL query to create the 'sleep' table
         create_sleep_table_query = '''
         CREATE TABLE IF NOT EXISTS sleep (
-            log_id INT AUTO_INCREMENT,
-            baby_id INT,
+            log_id BIGINT AUTO_INCREMENT,
+            baby_id BIGINT,
             sleep_duration INT,  -- Duration in hours
             log_date TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
             PRIMARY KEY (log_id)
         );
         '''
-
         # Execute the query to create the 'sleep' table
         cursor.execute(create_sleep_table_query)
         print("Table 'sleep' created successfully")
